Integrate the receiver app into the repository.
The receiver is implemented in form of an Android application that
uses the phone's camera and OpenCV to decode the transmission. This
patch integrates the app into the git repository so that both the
transmitter and the receiver can be built together and can depend
on a shared Java library.