Don't perform implicit PWM-GPIO conversion.
Before, the Pin object would automatically use setState method
if the value provided by upper layers is min or maximum. We no longer
do that, because we don't know the range of PWM values in advance.
It is better to let upper layers to perform the conversion explicitly
if necessary.