Nearest-color matching in demodulators based on RGB-space distance.
This patch implements nearest color matching in all FSK demodulators
based on the Euclidian distance between two colors in the
three-dimensional RGB space.

As part of the change, we also change the prototype of the demodulate
method back to the original type. The processing framework will give
us a Color object with detected hue and brightness attributes and with
the saturation attribute set to the maximum value. That makes it
possible to use the same type in all demodulators (both amplitude and
frequency).