Implement new type Amplitude for amplitude modulators.
The new type works just like Integer, but unlike Integer, an Euclidean
distance of two Amplitudes can be calculated.

Both Amplitude and Color types implement the EuclideanDistance
interface which means that they both can be used to obtain the nearest
neighbor in a space of modulation symbols.

Also update the UI to show the symbol detected by the modulator,
rather than the hue color detected by the processing framework.