Rename EuclideanSpace to Coordinate.
Coordinate is a better name for what this interface implements. The
Coordinate space requires each implementing class to provide an
euclidian distance method. It also provides a default implementation
for calculating the nearest neighbor out of a set of neighbors in the
coordinate space.

This patch also includes minor API improvements to the Amplitude class.