Add a LED driver for the text console.
This driver emulates an RGB led by changing the color of a rectangle
on the ANSI text console. The driver can only display the basic set of
colors: black, white, red, green, blue, yellow, cyan, and magenta. If
a different color is passed to it, it will display the nearest
matching color from that palette.