Make it possible to configure LED driver at runtime.
The system property led.driver can be used to specify the LED driver
to use, e.g., ConsoleLED or PiRgbLED. If the driver needs arguments to
be passed to the constructor, those can be provided via the property
led.params. The constructor must accept one string and parse it itself.