Fix the frame sampler.
The previous implementation did not work correctly. It would update
the timestamp of the last taken frame to the timestamp from the image,
which is not correct. This implementation works with a fixed timestamp
base and monotonically increases the base with given intervals,
regardless of the timestamp value of the incoming frame.