Implement Bus event handling in Processing class.
Rather than letting each block subscribe for the Bus event its needs,
we implement a setter in each block and let the Processing class to
handle all even registrations and subscriptions. That way, the
processing block does not need to know about the particular mechanism
used to deliver those events. Also, the processing class can correctly
unregister from the bus in its shutdown method.

To avoid dependencies between unrelated pieces of code, also define
most of the events in the Bus class itself so that they can be
accessed without the need to import unrelated pieces of code.