Improvements to RaptorQ.
Implement minPacketSize and maxPacketSize methods. The former returns
the minimum packet size that the consumer can expect. In case the
transmitted data is divisible by packet size, the function returns
packet size. The latter method always returns packet size.

Also, fix the calculation of the completion percentage which did not
work with the number of source symbols already received in the
previous version.