Split RaptorQ implementation into separate encoder and decoder.
This is necessary to implement a mode where the receiver does not know
the data it is receiving in advance. That will be useful, e.g., when
the transmission is activated by a button push on the IoT device,
without establishing a Wi-Fi P2P connection first. In that case, the
receiver will not necessarily know the data being transmitted in
advance.

This feature will alllow implementing advanced IoT scenarios where a
smartphone application is used to collect certificate fingerprint of
deployed devices even if the smartphone application cannot connect to
those devices via Wi-Fi P2P.