Do Hamming distance comparison in Viewfinder.
The Receiver class may not know what data we're receiving and will
thus not compare it. Instead, it sends the data it has received via
the Bus and whoever created the receiver object is in charge of
comparing it.

The Receiver class also implements start and stop methods. Start must
be called to begin transmission. stop will be called automatically
after the transmission has completed (and may also be called manually).