Proper Receiver object lifecycle management.
The receiver object provides an API which can be used to start and
stop the receiver, as well as to reset it. The reset method discards
any state and data acquired so far (transmission will start from
beginning).

Store the current receive object in the model so that it survives
device rotations and the activity going in and out of focus. Stop the
receiver when loosing focus and start it when focus is acquired again.

Reset and re-start the receiver after the user dismisses the dialog
showing that the transmission has completed.