Precise timing across transmitted frames.
The previous version of the code synchronized transmitted bits to wall
clock within a transmitted frame, but did not keep the synchronization
across frames. This version remembers the timestamp of the last
transmitted bit and the transmission of the first bit of the next
frame will be precisely timed with respect to the last bit of the
previous frame.

This change reduces error rate at higher transmission speeds.