Cleanup progress indicators.
We show two progress indicators, one showing the percentage of data
transferred, the other the percentage of data verified and decoded.

This patch performs some overdue cleanup of the code that calculates
and redraws those progress indicators.

Also, hide the live RoI preview since we only need that for
development purposes.