Convert LEDException into (unchecked) runtime exception.
The need to handle LEDException was causing too much headache across
the code base. In pretty much all cases, the code that handled
LEDException would turn it into a RuntimeException.

Since LEDException is in general generated in response to a low-level
hardware problem, it makes sense to turn it into an unchecked
exception and terminate the process if it occurs. Such low-level
exceptions are usually not recoverable anyway and indicate a serious
problem with the low-level subsystem or hardware.